:recycle: refactor: Implementação das alterações na listagem de vagas
