+++ src/components/ListarVagasMonitor.jsx
@@ -12,6 +12,7 @@ import CalcularValidade from "../util/CalcularValidade";
 import CalcularHoras from "../util/CalcularHoras";
 import LiberarVaga from "../util/LiberarVaga";
 import ValidarRequisicao from "../util/ValidarRequisicao";
+import { VagaMonitor } from "./VagaMonitor";
 
 const ListarVagasMonitor = () => {
   const [resposta, setResposta] = useState([]);
@@ -212,23 +213,6 @@ const ListarVagasMonitor = () => {
     });
   };
 
-  const funcLiberVaga = async (id_vaga, numero, index) => {
-    const requisicao = await createAPI();
-    requisicao.post(`/estacionamento/saida`, {
-        idvagaVeiculo: id_vaga,
-        vaga: numero,
-      }).then(async (response) => {
-        if (response.data.msg.resultado) {
-           await LiberarVaga(resposta, setResposta, index);
-        } else {
-           Swal.fire(`${response.data.msg.msg}`, "", "error");
-        }
-      })
-        .catch(function (error) {
-          ValidarRequisicao(error)
-      });
- }
-
 useEffect(() => {
   localStorage.removeItem("id_vagaveiculo");
     if (localStorage.getItem("numero_vaga")) {
@@ -257,16 +241,6 @@ useEffect(() => {
       setContador(contador + 1);
     }, 1000);
   }, [contador]);
-
-  const horaAgoraFunc = async () => {
-    const dataAtual = new Date();
-    const hora = dataAtual.getHours().toString().padStart(2, "0");
-    const minutos = dataAtual.getMinutes().toString().padStart(2, "0");
-    const segundos = dataAtual.getSeconds().toString().padStart(2, "0");
-    const horaAtual = `${hora}:${minutos}:${segundos}`;
-    setHoraAgora(horaAtual);
-    return horaAtual;
-  };
 
   useEffect(() => {
       (async () => {
@@ -314,298 +288,6 @@ useEffect(() => {
       localStorage.removeItem(`foto${i}`);
     }
   }, []);
-
-
-  const registroDebitoAutomatico = async (placa, numero, tempo, id_vaga, index) => {
-    const requisicao = await createAPI();
-
-    requisicao
-    .post("/estacionamento", {
-      placa: placa,
-      numero_vaga: numero,
-      tempo: tempo,
-      id_vaga_veiculo: id_vaga,
-    })
-    .then((response) => {
-      if (response.data.msg.resultado === true) {
-        ImpressaoTicketEstacionamento('PRIMEIRA',
-        response.data.data.chegada,
-        response.data.data.tempo,
-        response.config.headers.id_usuario,
-        [numero],
-        placa,
-        'debito',
-        '00:30:00',
-        response.data.data.notificacao_pendente
-        );
-
-        const updatedResposta = [...resposta];
-        let validade = CalcularHoras(CalcularValidade(response.data.data.chegada, response.data.data.tempo))
-        updatedResposta[index] = {
-          ...updatedResposta[index],
-          temporestante: CalcularValidade(response.data.data.chegada, response.data.data.tempo),
-          corline: validade.corline,
-          cor: validade.cor,
-        };
-        const newArray = updatedResposta.filter((item) => item !== undefined);
-        setResposta(newArray);
-        localStorage.setItem("listaVagas", JSON.stringify(newArray));
-
-      } else {
-        Swal.fire({
-          title: `${response.data.msg.msg}`,
-          showCancelButton: true,
-          showDenyButton: true,
-          cancelButtonText: "Cancelar",
-          confirmButtonText: "Liberar",
-          denyButtonText: `Notificar`,
-        }).then((result) => {
-          if (result.isConfirmed) {
-            funcLiberVaga(id_vaga, numero, index);
-          } else if (result.isDenied) {
-            localStorage.setItem("id_vagaveiculo", id_vaga);
-            localStorage.setItem("vaga", numero);
-            localStorage.setItem("placa", placa);
-            localStorage.setItem("idVagaVeiculo", id_vaga);
-            FuncTrocaComp("Notificacao");
-          }
-        });
-      }
-    })
-    .catch(function (error) {
-      ValidarRequisicao(error)
-    });
-
-
-  }
-
-  const funcExtratoPlaca = (placa) => {
-    const requisicao = createAPI();
-    requisicao.get(`/veiculo/${placa}`).then((response) => {
-        if ( response.data.msg.resultado === false 
-          && response.data.msg.msg !== "Dados encontrados"
-        ) {
-          setMensagem(response.data.msg.msg);
-          setTimeout(() => {
-            setEstado(false);
-            setMensagem("");
-          }, 3000);
-        } else {
-          const link = response?.data.data.map((item) => ({
-            placa: item.placa,
-            modelo: item.modelo.modelo,
-            fabricante: item.modelo.fabricante.fabricante,
-            cor: item.cor,
-            vaga: item.estacionado[0].numerovaga,
-            numero_notificacoes_pendentes: item.numero_notificacoes_pendentes,
-            saldo_devedor: item.saldo_devedorr,
-            estacionado: item.estacionado[0].estacionado,
-            tempo: item.estacionado[0].tempo,
-            chegada: item.estacionado[0].chegada,
-            temporestante: item.estacionado[0].temporestante,
-            id_vaga_veiculo: item.estacionado[0].id_vaga_veiculo,
-          }));
-
-          const impressao = link[0];
-
-          ImpressaoTicketEstacionamento(
-            "SEGUNDA",
-            impressao.chegada,
-            impressao.tempo,
-            "Nao informado",
-            impressao.vaga,
-            impressao.placa,
-            "Nao informado",
-            "Nao informado",
-            impressao.numero_notificacoes_pendentes
-          );
-        }
-      })
-      .catch((error) => {
-        if (
-          error?.response?.data?.msg === "Cabeçalho inválido!" ||
-          error?.response?.data?.msg === "Token inválido!" ||
-          error?.response?.data?.msg ===
-            "Usuário não possui o perfil mencionado!"
-        ) {
-          localStorage.removeItem("user");
-          localStorage.removeItem("token");
-          localStorage.removeItem("perfil");
-        } else {
-          console.log(error);
-        }
-      });
-  };
-
-  const estaciona = async (index) => {
-    localStorage.setItem("numero_vaga", resposta[index].numero);
-    const horaAgoraNew = await horaAgoraFunc();
-    if (resposta[index].temporestante < horaAgoraNew && resposta[index].placa !== "") {
-      if (resposta[index].numero_notificacoes_pendentes !== 0 || resposta[index].numero_notificacoes_pendentess  !== 0) {
-        if (horaAgoraNew < resposta[index].hora_notificacao) {
-          Swal.fire({
-            title: "Deseja liberar esta vaga?",
-            showCancelButton: true,
-            showDenyButton: true,
-            cancelButtonText: "Cancelar",
-            confirmButtonText: "Liberar",
-            denyButtonText: "Regularizar",
-            denyButtonColor: "#3A58C8",
-            footer: `
-            <button class="btn3 botao bg-blue-50" id="ticket">
-            Extrato de placa
-            </button>
-            `,
-          }).then((result) => {
-            if (result.isConfirmed) {
-              funcLiberVaga(resposta[index].id_vaga_veiculo, resposta[index].numero, index);
-            } else if (result.isDenied) {
-              localStorage.setItem("VagaVeiculoId", resposta[index].id_vaga_veiculo);
-              FuncTrocaComp("ListarNotificacoes");
-            }
-          });
-  
-          const btnFooter2 = document.getElementById("ticket");
-          btnFooter2.addEventListener("click", function () {
-            funcExtratoPlaca(resposta[index].placa);
-          });
-        } else {
-          Swal.fire({
-            title: "Deseja liberar esta vaga?",
-            showCancelButton: true,
-            showDenyButton: true,
-            cancelButtonText: "Cancelar",
-            confirmButtonText: "Liberar",
-            denyButtonText: `Notificar`,
-            footer: `
-            <button class="btn3 botao bg-blue-50" id="ticket">
-            Extrato de placa
-            </button>
-            `,
-          }).then((result) => {
-            if (result.isConfirmed) {
-              funcLiberVaga(resposta[index].id_vaga_veiculo, resposta[index].numero, index);
-            } else if (result.isDenied) {
-              localStorage.setItem("vaga", resposta[index].numero);
-              localStorage.setItem("placa", resposta[index].placa);
-              FuncTrocaComp("Notificacao");
-            }
-          });
-  
-          const btnFooter2 = document.getElementById("ticket");
-          btnFooter2.addEventListener("click", function () {
-            funcExtratoPlaca(resposta[index].placa);
-          });
-        }
-      } else if ( resposta[index].debito === "S") {
-        Swal.fire({
-          title: "Deseja liberar esta vaga?",
-          showDenyButton: true,
-          showCancelButton: true,
-          cancelButtonText: "Cancelar",
-          confirmButtonText: "Liberar",
-          denyButtonText: `Debitar`,
-          denyButtonColor: "green",
-          footer: `
-          <button class="btn3 botao bg-blue-50" id="ticket">
-          Extrato de placa
-          </button>
-          `,
-        }).then((result) => {
-          if (result.isConfirmed) {
-            funcLiberVaga(resposta[index].id_vaga_veiculo, resposta[index].numero, index);
-          } else if (result.isDenied) {
-            registroDebitoAutomatico(resposta[index].placa, resposta[index].numero, resposta[index].tempo, resposta[index].id_vaga_veiculo, index);
-          }
-        });
-
-        const btnFooter2 = document.getElementById("ticket");
-        btnFooter2.addEventListener("click", function () {
-          funcExtratoPlaca(resposta[index].placa);
-        });
-      } else {
-        Swal.fire({
-          title: "Deseja liberar esta vaga?",
-          showDenyButton: true,
-          showCancelButton: true,
-          cancelButtonText: "Cancelar",
-          confirmButtonText: "Liberar",
-          denyButtonText: "Notificar",
-          footer: `
-          <button class="btn3 botao bg-green-50 mx-2" id="btnFooter">
-          Adicionar tempo
-          </button>
-          <br />
-          <button class="btn3 botao bg-blue-50" id="ticket">
-          Extrato de placa
-          </button>
-          `,
-        }).then((result) => {
-          if (result.isConfirmed) {
-            funcLiberVaga(resposta[index].id_vaga_veiculo, resposta[index].numero, index);
-          } else if (result.isDenied) {
-            localStorage.setItem("id_vagaveiculo", resposta[index].id_vaga_veiculo);
-            localStorage.setItem("vaga", resposta[index].numero);
-            localStorage.setItem("placa", resposta[index].placa);
-            localStorage.setItem("idVagaVeiculo", resposta[index].id_vaga_veiculo);
-            FuncTrocaComp("Notificacao");
-          }
-        });
-
-        const btnFooter = document.getElementById("btnFooter");
-        btnFooter.addEventListener("click", function () {
-          localStorage.setItem("vaga", resposta[index].numero);
-          localStorage.setItem("id_vagaveiculo", resposta[index].id_vaga_veiculo);
-          localStorage.setItem("placa", resposta[index].placa);
-          localStorage.setItem("popup", true);
-          FuncTrocaComp("RegistrarVagaMonitor");
-          Swal.close();
-        });
-
-        const btnFooter2 = document.getElementById("ticket");
-        btnFooter2.addEventListener("click", function () {
-          funcExtratoPlaca(resposta[index].placa);
-        });
-      }
-    } else {
-      if ( resposta[index].placa === "") {
-        localStorage.setItem("vaga", resposta[index].numero);
-        localStorage.setItem("tipoVaga", resposta[index].tipo);
-        localStorage.setItem("popup", false);
-        FuncTrocaComp("RegistrarVagaMonitor");
-        return;
-      }
-        Swal.fire({
-          title: "Deseja liberar esse veículo?",
-          showDenyButton: true,
-          showCancelButton: true,
-          cancelButtonText: "Cancelar",
-          confirmButtonText: "Liberar",
-          denyButtonText: `Adicionar tempo`,
-          denyButtonColor: "green",
-          footer: `
-          <button class="btn3 botao bg-blue-50" id="ticket">
-          Extrato de placa
-          </button>
-          `,
-        }).then((result) => {
-          if (result.isConfirmed) {
-            funcLiberVaga(resposta[index].id_vaga_veiculo, resposta[index].numero, index);
-          } else if (result.isDenied) {
-            localStorage.setItem("vaga", resposta[index].numero);
-            localStorage.setItem("id_vagaveiculo", resposta[index].id_vaga_veiculo);
-            localStorage.setItem("placa", resposta[index].placa);
-            localStorage.setItem("popup", true);
-            FuncTrocaComp("RegistrarVagaMonitor");
-          }
-        });
-
-        const btnFooter2 = document.getElementById("ticket");
-        btnFooter2.addEventListener("click", function () {
-          funcExtratoPlaca(resposta[index].placa);
-        });
-    }
-  };
 
   return (
     <div className="dashboard-container mb-5">
@@ -721,68 +403,9 @@ useEffect(() => {
                       {resposta.length !== 0 ? (
                         resposta.map((vaga, index) => (
                         vaga !== null && (
-                        <tr
-                          key={index}
-                          className="card-list"
-                          data-vaga={vaga.numero}
-                          onClick={() => {
-                            estaciona(
-                              index
-                            );
-                          }}
-                        >
-                          <th
-                            className="text-white"
-                            scope="row"
-                            style={{
-                              backgroundColor: vaga.corvaga,
-                              color: vaga.cor,
-                            }}
-                          >
-                            {vaga.numero}
-                          </th>
-                          <td
-                            className="fw-bolder"
-                            style={{
-                              backgroundColor: vaga.corline,
-                              color: vaga.cor,
-                            }}
-                          >
-                            {vaga.placa == '0' ? null : vaga.placa}{" "}
-                            <small id={vaga.display}>
-                              {vaga.numero_notificacoes}
-                            </small>
-                          </td>
-                          <td
-                            className="fw-bolder"
-                            style={{
-                              backgroundColor: vaga.corline,
-                              color: vaga.cor,
-                            }}
-                          >
-                            {vaga.chegada}
-                          </td>
-                          <td
-                            className="fw-bolder"
-                            style={{
-                              backgroundColor: vaga.corline,
-                              color: vaga.cor,
-                            }}
-                          >
-                            <h6
-                              id={vaga.variaDisplay}
-                              style={{
-                                backgroundColor: vaga.corline,
-                                color: vaga.cor,
-                              }}
-                            >
-                              <span>{vaga.temporestante}</span>
-                            </h6>
-                          </td>
-                        </tr>
+                          <VagaMonitor vaga={vaga} index={index} key={index} setMensagem={setMensagem} setEstado={setEstado} setResposta={setResposta} resposta={resposta} />
                         )
                       ))) : ( null )}
-
                     </tbody>
                   </table>
                 </div>
